Fix bug install posh-git
The sudo command used to install posh git was malformed. The -NoProfile
switch was not working so the elevated powershell would try to install
the posh-git by elevating a powershell session, and this session would
try to install posh-git, and so on and so on. the result would be
multiple powershell windows and posh-git would not get installed.