Modified module install/importing
Now install modules using -scope currentuser so that we do not need to
be run elevated to install. Will still warn if not installing as -Scope
CurrentUser and not being run elevated that to install you need to run
with elevated permissions.
Removed some commented out code to change the background color if
running elevated. Now just sets the username/hostname to red.